requirements.txt
Django==3.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==2.2.2
whitenoise==5.2.0
dj-database-url==0.5.0
psycopg2==2.8.6
cloudinary==1.25.0
django-cloudinary-storage==0.3.0